requirements.txt
discord.py
deepl
Flask